pyproject.toml
[project]
name = "chartkit"
version = "0.1.0"
description = "Biblioteca de charting padronizado para DataFrames pandas"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "babel>=2.17.0",
    "cachetools>=6.2.6",
    "loguru>=0.7.3",
    "matplotlib>=3.10.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "pydantic-settings>=2.12.0",
]

[project.optional-dependencies]
jupyter = [
    "ipykernel>=6.29",
    "ipywidgets>=8.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chartkit"]

[tool.hatch.build.targets.sdist]
include = [
    "src/chartkit",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["-ra", "--strict-markers", "--strict-config", "--tb=short", "-q"]
markers = ["slow: marks slower tests (e.g. config loading with real files)"]
filterwarnings = ["error", "ignore::UserWarning:matplotlib"]

# Exemplo de configuracao do chartkit via pyproject.toml
# Descomente e customize conforme necessario:
#
# [tool.charting]
# [tool.charting.branding]
# company_name = "Minha Empresa"
# footer_format = "Fonte: {source}, {company_name}"
#
# [tool.charting.colors]
# primary = "#003366"
# secondary = "#0066CC"
#
# [tool.charting.layout]
# figsize = [12.0, 8.0]
# dpi = 150
